+++ examples/testing-with-bats/test.sh
@@ -1,1 +1,7 @@
-../../tests/test.sh
+#!/bin/bash
+
+# Get Packages
+./sbpl.sh
+
+# Run Tests
+./sbpl.sh test